swtpm: Add options for pluggable backend (--tpmstate backend)
Add two options to specify pluggable backend type and backend URI.

Ex:
  --tpmstate backend=...,backend_uri=...

Backend URI is specific to each backend plugin which points to the
location of the NVRAM files.
Currently, "file" is the only one available backend. In this case
backend_uri should be a path to the directory where files are stored.

This option is designed to compatible with existing "dir" option.
If "dir" is specified, swtpm prioritize "dir" ignoring "backend".

Signed-off-by: Eiichi Tsukata <eiichi.tsukata@nutanix.com>